Move service files with classes that interact with Common Platform into cp api folder
