+++ app/controllers/api/internal/v2/defendants_controller.rb
@@ -15,7 +15,7 @@ module Api
         end
 
         def show
-          defendant = DefendantFinder.call(defendant_id: params[:id])
+          defendant = CommonPlatformApi::DefendantFinder.call(defendant_id: params[:id])
 
           if defendant.present?
             render json: DefendantSerializer.new(defendant, serialization_options)
