+++ app/services/common_platform_api/record_laa_reference.rb
@@ -8,7 +8,7 @@ module CommonPlatformApi
                    status_code:,
                    application_reference:,
                    status_date:,
-                   connection: CommonPlatformConnection.call)
+                   connection: CommonPlatformApi::CommonPlatformConnection.call)
 
       @offence_id = offence_id
       @status_code = status_code
