+++ app/workers/representation_order_creator_worker.rb
@@ -5,7 +5,7 @@ class RepresentationOrderCreatorWorker
 
   def perform(request_id, defendant_id, offences, maat_reference, defence_organisation)
     Current.set(request_id: request_id) do
-      RepresentationOrderCreator.call(
+      CommonPlatformApi::RepresentationOrderCreator.call(
         defendant_id: defendant_id,
         offences: offences,
         maat_reference: maat_reference,
