+++ spec/services/common_platform_api/common_platform_connection_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe CommonPlatformConnection do
+RSpec.describe CommonPlatformApi::CommonPlatformConnection do
   subject(:connect_to_common_platform) { described_class.call(host: host) }
 
   let(:host) { "https://example.com" }
