+++ spec/services/common_platform_api/record_laa_reference_spec.rb
@@ -33,7 +33,7 @@ RSpec.describe CommonPlatformApi::RecordLaaReference do
   end
 
   context "with connection" do
-    let(:connection) { double("CommonPlatformConnection") }
+    let(:connection) { double("CommonPlatformApi::CommonPlatformConnection") }
     let(:request_params) do
       {
         statusCode: "ABCDEF",
