+++ spec/services/common_platform_api/record_representation_order_spec.rb
@@ -43,7 +43,7 @@ RSpec.describe CommonPlatformApi::RecordRepresentationOrder do
   end
 
   context "with connection" do
-    let(:connection) { double("CommonPlatformConnection") }
+    let(:connection) { double("CommonPlatformApi::CommonPlatformConnection") }
     let(:request_params) do
       {
         statusCode: "ABCDEF",
